test: stabilize metrics diagnostics and pause resume coverage
